+++ app/Http/Requests/Web/Account/UpdateAccountRequest.php
@@ -17,7 +17,7 @@ class UpdateAccountRequest extends FormRequest
             'ccp_number'          => ['sometimes', 'required', 'string', 'max:255', Rule::unique('accounts', 'ccp_number')->ignore($id)],
             'ccp_key'             => ['sometimes', 'required', 'string', 'max:255', new CcpKey],
             'draw_day'            => ['sometimes', 'required', 'integer', 'min:1', 'max:31'],
-            'min_withdraw_amount' => ['sometimes', 'required', 'integer', 'min:0'],
+            'min_withdraw_amount' => ['sometimes', 'required', 'numeric', 'min:0'],
             'max_withdraw_count'  => ['sometimes', 'required', 'integer', 'min:1'],
         ];
     }
