+++ app/Http/Requests/Web/Batch/UpdateBatchRequest.php
@@ -16,7 +16,7 @@ class UpdateBatchRequest extends FormRequest
         return [
             'source_id'        => ['nullable', 'integer'],
             'source_type'      => ['nullable', 'string', 'max:50'],
-            'purchase_price'   => ['sometimes', 'integer', 'min:0'],
+            'purchase_price'   => ['sometimes', 'numeric', 'min:0'],
             'initial_quantity' => ['sometimes', 'integer', 'min:0'],
             'current_quantity' => ['sometimes', 'integer', 'min:0'],
         ];
