+++ app/Http/Requests/Web/Stock/UpdateStockRequest.php
@@ -28,9 +28,9 @@ class UpdateStockRequest extends FormRequest
             'source_type'       => ['nullable', 'string'],
             'initial_quantity'  => ['nullable', 'integer', 'min:0'],
             'current_quantity'  => ['nullable', 'integer', 'min:0'],
-            'purchase_price'    => ['nullable', 'integer', 'min:0'],
-            'selling_price'     => ['nullable', 'integer', 'min:0'],
-            'installment_price' => ['nullable', 'integer', 'min:0'],
+            'purchase_price'    => ['nullable', 'numeric', 'min:0'],
+            'selling_price'     => ['nullable', 'numeric', 'min:0'],
+            'installment_price' => ['nullable', 'numeric', 'min:0'],
         ];
     }
 }
